+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - optimize weight of images used in rule descriptions
 - optimize Github actions workflows and CodeQL security corrections
 - deploy HTML files of rule content
+- add technology labels in rules file
 
 ### Deleted